add debugging
