+++ src/logging/debug/debug.h
@@ -1,0 +1,5 @@
+#pragma once 
+
+void initDebugging();
+void printDebugMessage(const char* debug_msg, debug_code_t debug_code);
+void logDebugCode(char* buf, debug_code_t debug_code);
