+++ src/logging/tasks.h
@@ -2,4 +2,6 @@
 
 void initSD();
 void initDataLogging(); 
+void initDebugging();
+
 void loopTaskLogData();
